Revert "Split up ExecutableImage into FB and SO versions (#1329)"
This reverts commit 452bcffe873964b6741b5773c75929aba30b846e.
