+++ inc/common/pjrt_implementation/loaded_executable_instance.h
@@ -8,35 +8,48 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 // https://llvm.org/LICENSE.txt
 
-#ifndef TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_LOADED_EXECUTABLE_INSTANCE_H_
-#define TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_LOADED_EXECUTABLE_INSTANCE_H_
-
 // c++ standard library includes
 #include <memory>
 #include <mutex>
+#include <optional>
+#include <string>
+#include <unordered_map>
+#include <unordered_set>
 #include <vector>
 
 // PJRT C API includes
 #include "xla/pjrt/c/pjrt_c_api.h"
 
+// tt-mlir includes
+#include "tt/runtime/runtime.h"
+
 // tt-xla includes
 #include "common/pjrt_implementation/device_instance.h"
-#include "common/pjrt_implementation/executable_image.h"
+#include "common/pjrt_implementation/executable_instance.h"
 #include "common/status.h"
 
+#ifndef TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_LOADED_EXECUTABLE_INSTANCE_H_
+#define TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_LOADED_EXECUTABLE_INSTANCE_H_
+
 namespace tt::pjrt {
+
+class BufferInstance;
+class ClientInstance;
 
 // Represents `PJRT_LoadedExecutable` structure and the functionality around it.
 // It is the in-memory loaded executable which is ready for input arguments to
 // execute.
 class LoadedExecutableInstance {
 public:
+  // Creates new loaded executable instance from the executable image.
+  static std::unique_ptr<LoadedExecutableInstance>
+  createInstance(std::shared_ptr<ExecutableImage> executable_image,
+                 std::vector<DeviceInstance *> &&addressable_devices,
+                 ClientInstance *client_instance);
+
   // Binds PJRT API functions implementation related to PJRT_LoadedExecutable
   // structure.
   static void bindApi(PJRT_Api *api);
-
-  // Virtual destructor for proper cleanup of derived classes
-  virtual ~LoadedExecutableInstance() = default;
 
   // Casts this loaded executable instance to PJRT_LoadedExecutable pointer.
   operator PJRT_LoadedExecutable *() {
@@ -67,9 +80,9 @@ public:
   void releaseResources();
 
   // Runs execution of this loaded executable.
-  virtual tt_pjrt_status execute(PJRT_LoadedExecutable_Execute_Args *args) = 0;
+  tt_pjrt_status execute(PJRT_LoadedExecutable_Execute_Args *args);
 
-protected:
+private:
   // Creates loaded executable instance from the executable image.
   LoadedExecutableInstance(
       std::shared_ptr<ExecutableImage> executable_image,
@@ -78,6 +91,71 @@ protected:
       : m_executable_image(std::move(executable_image)),
         m_addressable_devices(addressable_devices), m_deleted(false),
         m_client_instance(client_instance) {}
+
+  // Gets the appropriate runtime device from the `ClientInstance`.
+  std::optional<tt::runtime::Device>
+  getOrCreateMeshDevice(PJRT_Buffer *const *const *argument_lists,
+                        size_t num_args, size_t num_devices,
+                        PJRT_Device *pjrt_device);
+
+  // Collects device ids from the addressable devices.
+  std::unordered_set<int>
+  getDeviceIds(PJRT_Buffer *const *const *argument_lists, size_t num_args,
+               size_t num_devices);
+
+  // Gets input runtime tensors from the arguments' buffers and converts them to
+  // desired layout determined from the compiled graph.
+  tt_pjrt_status
+  getInputRuntimeTensors(PJRT_Buffer *const *const *argument_lists,
+                         size_t num_args, size_t num_devices,
+                         const tt::runtime::Device &runtime_device,
+                         std::uint32_t program_index,
+                         std::vector<tt::runtime::Tensor> &input_tensors);
+
+  // Returns an input tensor constructed from the provided buffer instances,
+  // prepared for execution. If we cannot reuse the already prepared tensor
+  // contained within the buffer instances, this will involve calling
+  // `toLayout()` which in most cases involves moving the data to the device.
+  std::optional<tt::runtime::Tensor>
+  prepareInputTensor(const std::vector<BufferInstance *> &arg_buffers,
+                     tt::runtime::Device device, size_t num_devices,
+                     std::uint32_t program_index, size_t arg_index);
+
+  // Fills strategy map from sharding configuration.
+  // TODO: This function might be better suited living in the tt-mlir
+  // repository. https://github.com/tenstorrent/tt-xla/issues/374
+  static mlir::FailureOr<std::unordered_map<std::string, std::string>>
+  fillStrategyMapFromSharding(
+      const mlir::tt::sharding_utils::MeshSharding &meshSharding,
+      size_t num_devices);
+
+  // Either returns single tensor or creates multi-device host tensor from arg
+  // tensors, depending on the strategy.
+  tt::runtime::Tensor getTensorFromStrategy(
+      const std::vector<BufferInstance *> &arg_buffers,
+      const std::unordered_map<std::string, std::string> &strategy);
+
+  // Converts input tensor to desired layout. This might move it on device.
+  tt::runtime::Tensor
+  convertTensorLayout(tt::runtime::Tensor input_tensor,
+                      std::uint32_t program_index, size_t arg_index,
+                      const tt::runtime::Device &runtime_device);
+
+  // Untilizes output tensors and transfers them from device to host.
+  tt_pjrt_status untilizeToHost(
+      const std::vector<tt::runtime::Tensor> &output_tensors,
+      size_t num_devices,
+      std::vector<std::vector<tt::runtime::Tensor>> &untilized_output_tensors);
+
+  // Fills the output lists of the PJRT API with the outputs of tt runtime
+  // execution.
+  void fillPJRTOutputLists(
+      const std::vector<std::vector<tt::runtime::Tensor>> &rt_outputs,
+      size_t num_devices, PJRT_Buffer **const *output_lists,
+      const std::vector<PJRT_Buffer_Type> &expected_output_data_types);
+
+  // Returns the shape of the output on the specified index.
+  std::vector<std::uint32_t> getOutputShape(size_t output_index);
 
   // Executable image instance which is shared between executable and loaded
   // executable instances.
