+++ inc/common/pjrt_implementation/module_builder/compile_options.h
@@ -2,27 +2,14 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_COMPILE_OPTIONS_H_
-#define TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_COMPILE_OPTIONS_H_
+#ifndef TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_MODULE_BUILDER_COMPILE_OPTIONS_H_
+#define TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_MODULE_BUILDER_COMPILE_OPTIONS_H_
 
 // c++ standard library includes
-#include <optional>
 #include <string>
 #include <unordered_map>
 
-namespace tt::pjrt {
-
-// Enumeration for backend types
-enum class BackendRuntime {
-  // Targets tt-mlir TTNN runtime to execute the compiled flatbuffer.
-  TTNNFlatbuffer,
-
-  // Generates TTNN C++ code to be compiled and executed.
-  TTNNCodegenCpp,
-
-  // Generates TTNN Python code.
-  TTNNCodegenPy
-};
+namespace tt::pjrt::module_builder {
 
 // POD struct containing various options used to customize module compilation.
 struct CompileOptions {
@@ -54,9 +41,6 @@ struct CompileOptions {
   // fixed.
   bool enable_fusing_conv2d_with_multiply_pattern = false;
 
-  // Backend runtime which should be targeted for compilation and execution.
-  BackendRuntime backend = BackendRuntime::TTNNFlatbuffer;
-
   static CompileOptions
   parse(const std::unordered_map<std::string, std::string> &compile_options);
 };
@@ -64,17 +48,12 @@ struct CompileOptions {
 namespace internal {
 
 // Parse out the value of one specific boolean flag from the options map.
-std::optional<bool> parseBoolOption(
-    const std::unordered_map<std::string, std::string> &compile_options,
-    std::string option_name);
-
-// Parse backend option from string to enum
-std::optional<BackendRuntime> parseBackendOption(
+bool parseBoolOption(
     const std::unordered_map<std::string, std::string> &compile_options,
     std::string option_name);
 
 } // namespace internal
 
-} // namespace tt::pjrt
+} // namespace tt::pjrt::module_builder
 
-#endif // TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_COMPILE_OPTIONS_H_
+#endif // TT_XLA_INC_COMMON_PJRT_IMPLEMENTATION_MODULE_BUILDER_COMPILE_OPTIONS_H_
