+++ src/common/CMakeLists.txt
@@ -41,10 +41,8 @@ add_library(TTPJRTCommon
     "pjrt_implementation/executable_image.cc"
     "pjrt_implementation/executable_instance.cc"
     "pjrt_implementation/loaded_executable_instance.cc"
-    "pjrt_implementation/flatbuffer_loaded_executable_instance.cc"
-    "pjrt_implementation/so_loaded_executable_instance.cc"
     "pjrt_implementation/memory_instance.cc"
-    "pjrt_implementation/compile_options.cc"
+    "pjrt_implementation/module_builder/compile_options.cc"
     "pjrt_implementation/module_builder/frontend_passes/shlo_input_role_propagation.cc"
     "pjrt_implementation/module_builder/module_builder.cc"
     "pjrt_implementation/serialized_executable_instance.cc"
